Applying La Trobe branding
The header and search tile have been styled to be more in line with La Trobe's branding (based on the uni web site & brand guide).

Only the default search has been updated so far (and not completely), so the database, newspaper, and journal searches will need to be updated as well.