Minor CSS updates
Added a style for the resource recommender.

Limited the hover style of drop-downs and inputs to the advanced search (so it doesn't affect the simple search elements).

The date is hidden from the brief results items for a database search (to match the dbases view).

Limited the drop-down in the Newspaper search's width to resolve a responsive issue when a newspaper with a long title is selected.