Update for the guided tour buttons
The guided tours are now launched from a separate button that appears in the top right (under the header). This button will animate dropping down when its tour changes (e.g. when moving from 'Library search' to 'Databases').

The guided tours are now changed when the '$locationChangeStart' event occurs, which happens when the URL is changed within Angular.

A separate guided tour has been added to the advanced search, which is launched from a '?' button next to its content.

Added a guided tour for 'My favourites'.