Menu & Guided tour update
The 3-dot menu's colours have been updated to better match LTU's styling.

When the guided tour is updated, it will now flag when the screen is being displayed in 'mobile view'. So the elements in the tour can be changed to suit. For some elements that are in the 3-dot menu for mobile view, that menu is triggered before they are highlighted (only done for the standard search so far).